Smaller file sizes
Now storing exponents to file instead of entire number! File size much down

+++ File_man.h
@@ -1,5 +1,6 @@
 #pragma once
 #include "options.h"
+#include <cstdint>
 #include <fstream>  // File RW
 #include <iostream> // For logging to console
 // Using cout, endl, and from std
@@ -12,9 +13,9 @@ public:
   // Function declarations
   void open_file(); // Creates or opens a file to store game data to
   int read_tile(int x, int y);
-  void write_tile(int x, int y, int val);
-  int read_score();
-  void write_score(int score);
+  void write_tile(int x, int y, uint8_t val);
+  uint64_t read_score();
+  void write_score(uint64_t score);
   bool is_new_file();
 
   std::fstream file;
